Code review feedback, switch ProblemDetailException to ProblemError.
