+++ tests/core/integration/test_settings.py
@@ -7,13 +7,13 @@ import pytest
 from pydantic import PositiveInt, validator
 from sqlalchemy.orm import Session
 
-from core.integration.exceptions import ProblemDetailException, SettingsValidationError
 from core.integration.settings import (
     BaseSettings,
     ConfigurationFormItem,
     ConfigurationFormItemType,
+    SettingsValidationError,
 )
-from core.util.problem_detail import ProblemDetail
+from core.util.problem_detail import ProblemDetail, ProblemError
 
 
 @pytest.fixture
@@ -79,9 +79,9 @@ def test_settings_init(mock_settings):
 
 
 def test_settings_init_invalid(mock_settings):
-    # Make sure that the settings class raises a ProblemDetailException
+    # Make sure that the settings class raises a ProblemError
     # when there is a problem with validation.
-    with pytest.raises(ProblemDetailException) as e:
+    with pytest.raises(ProblemError) as e:
         mock_settings(number=-1)
 
     problem_detail = e.value.problem_detail
@@ -91,7 +91,7 @@ def test_settings_init_invalid(mock_settings):
         == "'Number' validation error: ensure this value is greater than 0."
     )
 
-    with pytest.raises(ProblemDetailException) as e:
+    with pytest.raises(ProblemError) as e:
         mock_settings()
 
     problem_detail = e.value.problem_detail
@@ -112,9 +112,9 @@ def test_settings_validation(mock_settings):
 
 def test_settings_validation_custom(mock_settings, mock_problem_detail):
     # We can also add custom validation functions to the settings class.
-    # These functions should raise a ProblemDetailException if there is
+    # These functions should raise a ProblemError if there is
     # a problem with validation.
-    with pytest.raises(ProblemDetailException) as e:
+    with pytest.raises(ProblemError) as e:
         mock_settings(number=1, test="xyz")
 
     problem_detail = e.value.problem_detail
